+++ src/patch/core/SharedDef.zig
@@ -73,7 +73,7 @@ pub const GlobalState = extern struct {
     } = .{},
 };
 
-pub const GLOBAL_FUNCTION_VERSION = 22;
+pub const GLOBAL_FUNCTION_VERSION = 23;
 
 pub const GlobalFunction = extern struct {
     // Settings
@@ -94,6 +94,7 @@ pub const GlobalFunction = extern struct {
     GDrawText: *const fn (layer: GDrawLayer, text: ?*TextDef) bool,
     //GDrawTextBox: *const fn (layer: GDrawLayer, text: ?*TextDef, pad_x: i16, pad_y: i16, rect_color: u32) bool,
     GDrawRect: *const fn (layer: GDrawLayer, x: i16, y: i16, w: i16, h: i16, color: u32) bool,
+    GDrawRectBdr: *const fn (layer: GDrawLayer, x: i16, y: i16, w: i16, h: i16, color: u32, bdr_w: i16, bdr_col: u32) bool,
     GFreezeOn: *const fn () bool,
     GFreezeOff: *const fn () bool,
     GFreezeIsOn: *const fn () bool,